refactor(macos): split monolithic lib.rs into focused modules
Split nokhwa-bindings-macos/src/lib.rs (~2,422 lines) into:
- ffi.rs: FFI declarations, extern C blocks, framework links, type aliases
- format.rs: Format conversion helpers (str/nsstr, array conversions, FourCC)
- callback.rs: ObjC callback class registration and AVCaptureVideoCallback
- device.rs: AVCaptureDevice, discovery session, enums, device types
- session.rs: AVCaptureDeviceInput, AVCaptureVideoDataOutput, AVCaptureSession

Pure structural refactoring — no logic changes. Public API surface preserved.
